requirements.txt
python-telegram-bot==21.3
APScheduler==3.11.0
python-dotenv==1.0.0
schedule==1.2.2
aiosqlite==0.19.0